Fix circularity issues
