+++ packages/api-client/package.json
@@ -11,7 +11,7 @@
     "access": "public"
   },
   "dependencies": {
-    "@tryrolljs/sdk": "^0.1.2",
+    "@tryrolljs/sdk": "^0.1.4",
     "axios": "^1.2.2",
     "better-queue": "^3.8.12",
     "better-queue-memory": "^1.0.4",
@@ -22,7 +22,7 @@
     "@rollup/plugin-commonjs": "^22.0.2",
     "@rollup/plugin-typescript": "^11.0.0",
     "@swc/core": "^1.3.1",
-    "@swc/jest": "^0.2.22",
+    "@swc/jest": "^0.2.23",
     "@types/better-queue": "^3.8.3",
     "@types/jest": "^29.0.2",
     "@types/qs": "^6.9.7",
