+++ packages/api/package.json
@@ -11,20 +11,18 @@
     "access": "public"
   },
   "dependencies": {
-    "@tryrolljs/api-client": "0.0.1",
+    "@tryrolljs/api-client": "^0.0.1",
     "axios": "^1.2.2",
-    "events": "^3.3.0",
-    "qs": "^6.11.0"
+    "events": "^3.3.0"
   },
   "devDependencies": {
     "@rollup/plugin-babel": "^5.3.1",
     "@rollup/plugin-commonjs": "^22.0.2",
     "@rollup/plugin-typescript": "^11.0.0",
     "@swc/core": "^1.3.1",
-    "@swc/jest": "^0.2.22",
+    "@swc/jest": "^0.2.23",
     "@types/better-queue": "^3.8.3",
-    "@types/jest": "^29.0.2",
-    "@types/qs": "^6.9.7",
+    "@types/jest": "^29.2.0",
     "babel-jest": "^29.3.1",
     "jest": "^29.0.3",
     "jest-environment-jsdom": "^29.0.3",
