--- packages/api/src/auth/index.ts
@@ -1,2 +1,0 @@
-export * as types from './types'
-export * from './auth'
