+++ packages/api/src/index.ts
@@ -1,3 +1,2 @@
-export * as auth from './auth'
 export * as transaction from './transaction'
 export * as user from './user'
