--- packages/api/src/utils/get-random-string.ts
@@ -1,8 +1,0 @@
-const getRandomString = () => {
-  return (
-    Math.random().toString(36).substring(2, 15) +
-    Math.random().toString(36).substring(2, 15)
-  )
-}
-
-export default getRandomString
