--- packages/api/src/utils/index.ts
@@ -1,2 +1,0 @@
-export { default as getRandomString } from './get-random-string'
-export { default as pickQueryStringItemFromUrl } from './pick-query-string-item-from-url'
