+++ packages/feature-flag/package.json
@@ -21,7 +21,7 @@
     "@rollup/plugin-commonjs": "^22.0.2",
     "@rollup/plugin-typescript": "^11.0.0",
     "@swc/core": "^1.3.1",
-    "@swc/jest": "^0.2.22",
+    "@swc/jest": "^0.2.23",
     "@testing-library/react": "^13.4.0",
     "@types/jest": "^29.0.2",
     "@types/lodash.groupby": "^4.6.7",
