+++ packages/sdk/package.json
@@ -11,8 +11,8 @@
     "access": "public"
   },
   "dependencies": {
-    "@tryrolljs/api": "0.1.2",
     "event-emitter": "^0.3.5",
+    "qs": "^6.11.0",
     "type-fest": "^3.1.0"
   },
   "devDependencies": {
@@ -20,8 +20,9 @@
     "@rollup/plugin-commonjs": "^22.0.2",
     "@rollup/plugin-typescript": "^11.0.0",
     "@swc/core": "^1.3.1",
-    "@swc/jest": "^0.2.22",
+    "@swc/jest": "^0.2.23",
     "@types/jest": "^29.0.2",
+    "@types/qs": "^6.9.7",
     "jest": "^29.0.3",
     "jest-environment-jsdom": "^29.0.3",
     "rollup": "^2.78.1",
