+++ packages/sdk/src/auth/utils.ts
@@ -11,3 +11,10 @@ export const isLastUpdateTimestampExpired = (
 
   return now.getTime() - lastUpdateTimestamp >= expirationInMiliseconds
 }
+
+export const getRandomString = () => {
+  return (
+    Math.random().toString(36).substring(2, 15) +
+    Math.random().toString(36).substring(2, 15)
+  )
+}
